+++ src/components/FeaturedProductsGrid/FeaturedProductsGrid.jsx
@@ -1,15 +1,7 @@
-import styled from 'styled-components';
+
 import PropTypes from 'prop-types';
 import Product from './Product/Product';
-const Wrapper = styled.div`
-  width: 90%;
-  margin: auto;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-gap: 10px;
-  grid-auto-rows: minmax(100px, auto);
-  margin-bottom: 2%;
-`;
+import Wrapper from '../../styles/FeatureProducsGridWrapper.styled';
 
 const FeaturedProductGrid = ({ products }) => {
   return (
