+++ src/components/Footer/Footer.jsx
@@ -1,16 +1,4 @@
-import styled from 'styled-components';
-
-const FooterDiv = styled.footer`
-    height: 3rem;
-    background-color: black;
-    color: white;
-    position: relative;
-    width: 100%;
-    bottom: 0;
-    color: white;
-    font-size: 1em; 
-    text-align: center;
-`;
+import FooterDiv from "../../styles/FooterDiv.styled";
 
 const Footer = () => (
   <FooterDiv>
