+++ src/components/ProductCategories/ProductCategories.jsx
@@ -1,23 +1,7 @@
-import styled from 'styled-components';
 import ProductCategory from './ProductCategory/ProductCategory';
 import PropTypes from 'prop-types';
-const Wrapper = styled.div`
-  width: 90%;
-  margin: auto;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-gap: 10px;
-  grid-auto-rows: minmax(100px, auto);
-  margin-bottom: 2%;
-`;
-
-const Title = styled.h3`
-  font-size: 4vh;
-  @media (max-width: 420px) {
-    display: none;
-  }
-`;
-
+import Wrapper from '../../styles/ProductCategoriesWrapper.styled';
+import Title from '../../styles/CategoriesTitle.styled';
 
 const ProductCategories = ({ categories }) => {
   return (
