+++ src/components/ProductCategories/ProductCategory/ProductCategory.jsx
@@ -1,17 +1,7 @@
-import styled from 'styled-components';
 import PropTypes from 'prop-types';
-const Cell = styled.div`
-  display: inline-table;
-  background: #ff9d9c;
-  text-align: center;
-  border: black 5px solid;
-  font-size: 2.8vw;
-  color: black;
-`;
+import Cell from '../../../styles/ProductCategoryCell.styled';
+import CategoryImage from '../../../styles/CategoryImage.styled';
 
-const CategoryImage = styled.img`
-    max-width: 100%;
-}`;
 const ProductCategory = ({ category }) => {
   return (
     <Cell>
