+++ src/styles/Cart.styled.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+const Cart = styled.img`
+  text-align: right;
+  padding: 1%;
+  height: 6vh;
+  width: 70%;
+`;
+export default Cart;
