+++ src/styles/CategoriesTitle.styled.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const Title = styled.h3`
+  font-size: 4vh;
+  @media (max-width: 420px) {
+    display: none;
+  }
+`;
+export default Title;
