+++ src/styles/CategoryImage.styled.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+const CategoryImage = styled.img`
+max-width: 100%;
+}`;
+export default CategoryImage;
