+++ src/styles/Cell.styled.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const Cell = styled.div`
+  display: inline-table;
+  background: grey;
+  text-align: center;
+  border: black 5px solid;
+  heigth: 20%;
+  position: relative;
+`;
+
+export default Cell;
