+++ src/styles/FooterDiv.styled.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+const FooterDiv = styled.footer`
+  height: 3rem;
+  background-color: black;
+  color: white;
+  position: relative;
+  width: 100%;
+  bottom: 0;
+  color: white;
+  font-size: 1em;
+  text-align: center;
+`;
+
+export default FooterDiv;
