+++ src/styles/HeaderWrapper.styled.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const Wrapper = styled.div`
+  background-color: black;
+  height: 8vh;
+  color: white;
+  width: 100%;
+`;
+
+export default Wrapper;
