+++ src/styles/Logo.styled.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const Logo = styled.img`
+  float: left;
+  height: 6vh;
+`;
+
+export default Logo;
