+++ src/styles/NavLinkIcons.styled.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+import { Link } from 'react-router-dom';
+
+const NavLinkIcons = styled(Link)`
+  float: left;
+  display: flex;
+  color: #f2f2f2;
+  text-align: center;
+  padding-top: 5%;
+  text-decoration: none;
+  font-size: 15px;
+  padding-top: 8px;
+`;
+
+export default NavLinkIcons;
