+++ src/styles/PillsWrapper.styled.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const PillsWrapper = styled.div`
+  position: absolute;
+  width: 100%;
+  bottom: 0;
+  color: white;
+  font-size: 1rem;
+  text-align: center;
+`;
+
+export default PillsWrapper;
