+++ src/styles/ProducImage.styled.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+const ProductImage = styled.img`
+max-width: 100%;
+}`;
+
+export default ProductImage;
