+++ src/styles/ProductCategoriesWrapper.styled.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+const Wrapper = styled.div`
+  width: 90%;
+  margin: auto;
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-gap: 10px;
+  grid-auto-rows: minmax(100px, auto);
+  margin-bottom: 2%;
+`;
+
+export default Wrapper;
