+++ src/styles/ProductCategoryCell.styled.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+const Cell = styled.div`
+  display: inline-table;
+  background: #ff9d9c;
+  text-align: center;
+  border: black 5px solid;
+  font-size: 2.8vw;
+  color: black;
+`;
+export default Cell;
