+++ src/styles/RigthWrapper.styled.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+const RigthWrapper = styled.div`
+  float: right;
+  padding: 0;
+  margin: 0;
+`;
+
+export default RigthWrapper;
