+++ src/styles/SearchIcon.styled.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+const Search = styled.img`
+  margin: 0;
+  padding: 0;
+  height: 6vh;
+  width: 70%;
+`;
+
+export default Search;
